Version 2.9c
Version 2.9c

+++ src/main/java/nc/fluid/FluidFission.java
@@ -4,22 +4,22 @@ import nc.util.ColorHelper;
 
 public class FluidFission extends FluidMolten {
 	
-	private static final int FLUORINE_COLOR = 0xD3C85D;
-	private static final int FLIBE_COLOR = 0xC1C8B0;
+	private static final Integer FLUORINE_COLOR = 0xD3C85D;
+	private static final Integer FLIBE_COLOR = 0xC1C8B0;
 			
 	public FluidFission(String fluidName) {
 		super(fluidName);
 	}
 	
-	public FluidFission(String fluidName, int colour) {
+	public FluidFission(String fluidName, Integer colour) {
 		super(fluidName, colour);
 	}
 	
-	public static int getFluorideColor(int colour) {
-		return ColorHelper.blend(colour, FLUORINE_COLOR, 0.125F);
+	public static int getFluorideColor(Integer colour) {
+		return ColorHelper.blend(colour.intValue(), FLUORINE_COLOR, 0.125F);
 	}
 	
-	public static int getFLIBEColor(int colour) {
-		return ColorHelper.blend(colour, FLIBE_COLOR, 0.4F);
+	public static int getFLIBEColor(Integer colour) {
+		return ColorHelper.blend(colour.intValue(), FLIBE_COLOR, 0.4F);
 	}
 }
